added filters to catalog page and got them working

+++ WebShop_NULL/Infrastructure/Filters/FilterDTOs/DecimalFilterDto.cs
@@ -7,8 +7,8 @@ namespace WebShop_NULL.Infrastructure.Filters.FilterDTOs
     [ForPropertyType(PropertyTypeEnum.Decimal)]
     public class DecimalFilterDto : FilterDTO
     {
-        public decimal Min { get; set; } = Decimal.MinValue;
-        public decimal Max { get; set; } = Decimal.MaxValue;
+        public decimal Min { get; set; } = long.MinValue;
+        public decimal Max { get; set; } = long.MaxValue;
         
         protected override Expression<Func<Product, bool>> GenerateExpression()
         {
