+++ WebShop_NULL/Infrastructure/Filters/FilterDTOs/IntegerFilterDto.cs
@@ -7,8 +7,8 @@ namespace WebShop_NULL.Infrastructure.Filters.FilterDTOs
     [ForPropertyType(PropertyTypeEnum.Integer)]
     public class IntegerFilterDto : FilterDTO
     {
-        public int Min { get; set; } = Int32.MinValue;
-        public int Max { get; set; } = Int32.MaxValue;
+        public long Min { get; set; } = long.MinValue;
+        public long Max { get; set; } = long.MaxValue;
         
         protected override Expression<Func<Product, bool>> GenerateExpression()
         {
