+++ WebShop_NULL/Models/ViewModels/FilterViewModels/IntegerFilterViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using DomainModels;
+using WebShop_FSharp.ViewModels.CatalogModels;
 using WebShop_NULL.Infrastructure.Filters;
 
 namespace WebShop_NULL.Models.ViewModels.FilterViewModels
@@ -6,7 +7,7 @@ namespace WebShop_NULL.Models.ViewModels.FilterViewModels
     [ForPropertyType(PropertyTypeEnum.Integer)]
     public class IntegerFilterViewModel : FilterViewModel
     {
-        public int Min { get; set; } = 0;
-        public int Max { get; set; } = 0;
+        public long Min { get; set; } = 0;
+        public long Max { get; set; } = 0;
     }
 }
